Extract NonEmptyCollectionValidator from NameableSet and add corresponding helper method
