+++ server/webapp/WEB-INF/rails/webpack/models/mixins/new_validatable_mixin_spec.ts
@@ -23,6 +23,64 @@ import {ErrorMessages} from "models/mixins/validatable";
 import * as s from "underscore.string";
 
 describe("Validatable", () => {
+  describe("validateNonEmptyCollection", () => {
+    describe("validates that Iterables are not empty:", () => {
+      it("String", () => {
+        const t = new Thang();
+
+        expect(t.validate("string").hasErrors("string")).toBe(true);
+        expect(t.validate("string").errorsForDisplay("string")).toBe("this string is empty.");
+        t.string = "foo";
+        expect(t.validate("string").hasErrors("string")).toBe(false);
+        expect(t.validate("string").errorsForDisplay("string")).toBe("");
+      });
+
+      it("Array", () => {
+        const t = new Thang();
+
+        expect(t.validate("array").hasErrors("array")).toBe(true);
+        expect(t.validate("array").errorsForDisplay("array")).toBe("this array is empty.");
+        t.array = [1];
+        expect(t.validate("array").hasErrors("array")).toBe(false);
+        expect(t.validate("array").errorsForDisplay("array")).toBe("");
+      });
+
+      it("Map", () => {
+        const t = new Thang();
+
+        expect(t.validate("map").hasErrors("map")).toBe(true);
+        expect(t.validate("map").errorsForDisplay("map")).toBe("this map is empty.");
+        t.map = new Map<string, number>().set("a", 1);
+        expect(t.validate("map").hasErrors("map")).toBe(false);
+        expect(t.validate("map").errorsForDisplay("map")).toBe("");
+      });
+
+      it("Set", () => {
+        const t = new Thang();
+
+        expect(t.validate("set").hasErrors("set")).toBe(true);
+        expect(t.validate("set").errorsForDisplay("set")).toBe("this set is empty.");
+        t.set = new Set<number>([1]);
+        expect(t.validate("set").hasErrors("set")).toBe(false);
+        expect(t.validate("set").errorsForDisplay("set")).toBe("");
+      });
+    });
+  });
+
+  class Thang extends ValidatableMixin {
+    string: string = "";
+    map: Map<string, number> = new Map<string, number>();
+    array: number[] = [];
+    set: Set<number> = new Set<number>();
+
+    constructor() {
+      super();
+      this.validateNonEmptyCollection("string", {message: "this string is empty"});
+      this.validateNonEmptyCollection("array", {message: "this array is empty"});
+      this.validateNonEmptyCollection("map", {message: "this map is empty"});
+      this.validateNonEmptyCollection("set", {message: "this set is empty"});
+    }
+  }
 
   describe("validatePresenceOf", () => {
